Function... that does something!
this creates files that have the headlines listed.. but not correctly
or in the format we need.

but the code is also pulling 10 headlines for every jan. 1st from
1900-2000

it was also writing out to a JSON file, but w the function in there
that doesn’t seem to work anymore?